Cambios para poner en ambiente de pruebas

+++ database/migrations/2020_06_18_224915_create_product_categories_table.php
@@ -15,7 +15,7 @@ class CreateProductCategoriesTable extends Migration{
             $table->smallIncrements('id');
             $table->string('name', 50);
             $table->tinyInteger('deep');
-            $table->tinyInteger('root');
+            $table->smallInteger('root');
         });
     }
